+++ storage/sqlite/migrations/000024_review_session_questionnaire.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE review_sessions DROP COLUMN questionnaire;